requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
tavily-python>=0.5.0
